--- src/account/account.service.spec.ts
@@ -1,5 +1,0 @@
-import { AccountService } from './checking-account.service';
-import { AccountFactory } from './factories/account.factory';
-
-const accountService = new AccountService();
-const accountFactory = new AccountFactory();
