+++ src/account/application/saving-account.module.ts
@@ -1,10 +1,10 @@
 import { Module } from '@nestjs/common';
-import { SavingAccountController } from './saving-account.controller';
+import { SavingAccountController } from '../presenter/http/saving-account.controller';
 import { SavingAccountService } from './saving-account.service';
 import { AccountService } from './checking-account.service';
 
 @Module({
   controllers: [SavingAccountController],
-  providers: [AccountService, SavingAccountService]
+  providers: [AccountService, SavingAccountService],
 })
 export class SavingAccountModule {}
