+++ src/app.module.ts
@@ -2,9 +2,9 @@ import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
 import { ClientAccountModule } from './person/client.module';
-import { AccountModule } from './account/account.module';
-import { ManagerModule } from './person/manager.module';
-import { SavingAccountModule } from './account/saving-account.module';
+import { AccountModule } from './account/application/account.module';
+import { ManagerModule } from './person/application/manager.module';
+import { SavingAccountModule } from './account/application/saving-account.module';
 import { PostalServiceModule } from './postal-service/postal-service.module';
 
 @Module({
