+++ src/person/adapters/data/manager.json
