+++ src/person/domain/manager-model.ts
